+++ src/AuthContext.jsx
@@ -16,7 +16,7 @@ export const AuthProvider = ({ children }) => {
   useEffect(() => {
     const checkAuth = async () => {
       try {
-        const response = await fetchWithAuth("https://projekti2025backend-e0dubhd7e5h6akcw.swedencentral-01.azurewebsites.net/api/profile/", {
+        const response = await fetchWithAuth("http://127.0.0.1:8000/api/profile/", {
           method: "GET",
           credentials: "include",
         });
